requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
jinja2==3.1.3
celery==5.4.0
redis==5.0.3
sqlalchemy==2.0.27
pandas==2.2.1
pyyaml==6.0.1
openai-whisper==20231117  # Local Whisper; consider faster-whisper for optimization
slowapi==0.1.9
python-multipart==0.0.9
structlog==24.1.0
gunicorn==22.0.0
pytest==8.1.1
